+++ includes/class-node-names.php
@@ -22,4 +22,5 @@ final class Node_Names {
 	public const SSE                 = '_sse';
 	public const METADATA            = '_metadata';
 	public const UPTIME              = '_uptime';
+	public const COMPLETION          = '_completion';
 }